Drop cgroups config types from linux feature flag
